--- biotaphy/analyses/ancestral_state/__init__.py
@@ -1,1 +1,0 @@
-"""Ancestral state module __init__."""
