--- biotaphy/analyses/helpers/__init__.py
@@ -1,1 +1,0 @@
-"""Helpers module __init__."""
