--- biotaphy/analyses/pam_stats/__init__.py
@@ -1,1 +1,0 @@
-"""PAM stats module __init__."""
