--- biotaphy/analyses/phylo_beta_diversity/__init__.py
@@ -1,1 +1,0 @@
-"""Phylogenetic beta diversity module __init__."""
