--- biotaphy/analyses/tools/__init__.py
@@ -1,1 +1,0 @@
-"""Analyses tools module __init__."""
