+++ biotaphy/common/__init__.py
@@ -1,0 +1,1 @@
+"""Common module __init__."""
