+++ biotaphy/tools/ancestral_distribution.py
@@ -9,10 +9,10 @@ import argparse
 
 from lmpy import TreeWrapper
 
-from biotaphy.analyses.ancestral_state import anc_dp
-import biotaphy.analyses.tools.annotators as annotators
-import biotaphy.analyses.tools.plots as tree_plots
-from biotaphy.analyses.helpers import data_readers
+from biotaphy.analyses import anc_dp
+import biotaphy.common.annotators as annotators
+import biotaphy.common.plots as tree_plots
+from biotaphy.common import data_readers
 
 DESCRIPTION = """\
 Generates ancestral distribution estimations based on the environmental
