+++ tests/test_analyses/test_anc_dp.py
@@ -12,8 +12,8 @@ import pytest
 
 from lmpy import TreeWrapper
 
-from biotaphy.analyses.ancestral_state import anc_dp
-import biotaphy.analyses.helpers.data_readers as data_readers
+from biotaphy.analyses import anc_dp
+import biotaphy.common.data_readers as data_readers
 
 
 # .....................................................................................
