--- tests/test_analyses/test_ancestral_state/__init__.py
@@ -1,1 +1,0 @@
-"""Test ancestral state module __init__."""
