--- tests/test_analyses/test_helpers/__init__.py
@@ -1,1 +1,0 @@
-"""Test helpers module __init__."""
