--- tests/test_analyses/test_pam_stats/__init__.py
@@ -1,1 +1,0 @@
-"""Test PAM stats module __init__."""
