--- tests/test_analyses/test_phylo_beta_diversity/__init__.py
@@ -1,1 +1,0 @@
-"""Test phylo beta diversity module __init__."""
