+++ tests/test_analyses/test_site_statistics.py
@@ -3,13 +3,21 @@ import numpy as np
 
 from lmpy import Matrix, TreeWrapper
 
-from biotaphy.analyses.pam_stats.site_statistics import (
-    get_tip_lengths, mean_node_height, median_node_height,
-    node_height_percentile_2_5, node_height_percentile_25,
-    node_height_percentile_75, node_height_percentile_97_5,
-    mean_tip_length, median_tip_length, tip_length_percentile_2_5,
-    tip_length_percentile_25, tip_length_percentile_75,
-    tip_length_percentile_97_5, calculate_tree_site_statistics
+from biotaphy.analyses.site_statistics import (
+    get_tip_lengths,
+    mean_node_height,
+    median_node_height,
+    node_height_percentile_2_5,
+    node_height_percentile_25,
+    node_height_percentile_75,
+    node_height_percentile_97_5,
+    mean_tip_length,
+    median_tip_length,
+    tip_length_percentile_2_5,
+    tip_length_percentile_25,
+    tip_length_percentile_75,
+    tip_length_percentile_97_5,
+    calculate_tree_site_statistics,
 )
