+++ tests/test_common/__init__.py
@@ -1,0 +1,1 @@
+"""Tests common module."""
